split transmitter module
